parsing - reorganize
