+++ src/utility/index.ts
@@ -1,7 +1,7 @@
-import { CSVItemParser } from "./CSVItemParser";
-import { CSVTemplateParser } from "./CSVTemplateParser";
-import { FileLoader } from "./FileLoader";
-import { ParserManager } from "./ParserManager";
+import { CSVItemParser } from "./parsing/CSVItemParser";
+import { CSVTemplateParser } from "./parsing/CSVTemplateParser";
+import { FileLoader } from "./file/FileLoader";
+import { ParserManager } from "./parsing/ParserManager";
 
 export function prod() {
     return ParserManager;
