+++ src/utility/parsing/Parser.ts
@@ -1,5 +1,5 @@
-import type { IBasicItem } from "../models/basicItem";
-import type { ITemplate } from "../models/template";
+import type { IBasicItem } from "../../models/basicItem";
+import type { ITemplate } from "../../models/template";
 
 export interface Parser {
     parse(input:String) : void;
